Update classes for moved AppDataModels
